+++ tests/test_path_directives.py
